move ensureDefined function
