+++ src/app/modules/group/components/group-composition-filter/group-composition-filter.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { ensureDefined } from 'src/app/shared/helpers/null-undefined-predicates';
+import { ensureDefined } from 'src/app/shared/helpers/assert';
 
 export enum TypeFilter {
   Groups = 'groups',
