+++ src/app/modules/shared-components/components/selection/selection.component.ts
@@ -1,7 +1,7 @@
 
 import { Component, Input, Output, EventEmitter, OnChanges, SimpleChanges, forwardRef } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { ensureDefined } from 'src/app/shared/helpers/null-undefined-predicates';
+import { ensureDefined } from 'src/app/shared/helpers/assert';
 
 /**
  * To use inside form, just set the formControlName
